Update grafana config for apps with logstasher > 1
This is not exhaustive, but it's the ones that I know about at present.

Removing the field prefix means that we look for (for example) the `duration`
field in Kibana logs rather than `@fields.duration`.  This changes when we use
a version of logstasher that is greater than 1.  This is automatically used if
we've upgraded to govuk_app_config v1 or higher.