Update deploy Puppet job to deploy to both envs
The Carrenza Integration environment will still exist, but mostly just
for CI. This means we're going to keep a bunch of the management
structure in place, including Graphite, Icinga and Deploy Jenkins. To
ensure we continue to push the latest Puppet code to the Deploy Jenkins,
we need to update the alias assigned to the hosts file in Carrenza. This
will also allow the CI Jenkins to appropriately find the correct live
Integration stack doing a DNS look-up.