Use the Xenial HWE stack on Trusty machines
Install the Xenial kernel on Trusty (14.04 LTS) machines in order to
get the latest kernel updates, including for Spectre/Meltdown.

Change the icinga check script to check for this kernel version, and
remove the ability to check our precise machines, for simplicity.

We only run precise on two classes of machines in production -
puppetmaster and licensify-lb. We need to patch these manually for now.

See https://wiki.ubuntu.com/Kernel/LTSEnablementStack