Configure Deploy App Jenkins job for Docker tagging
This binds the Docker Hub credentials show they can be used in the
deploy job build. It adds the text finder plugin so that any builds
which output "DOCKER TAG FAILED" can be marked as unstable. Finally, it
changes the configuration of the Smokey job to run even if the build is
marked as unstable.

Paired with this is the need to set up the govukci-docker-hub credential
ID on each machine.