AWS set proxy_pass asset_manager_host with https
On AWS, static needs to proxy requests to asset_manager using
https, because it's the only port we want to enable on the ELB.

The nginx_enable_ssl variable that the static vhost uses to
set the protocol is disabled on AWS, because it's used in other
places to set the vhost listening port. We add a new variable
to evaluate the conditional instead of playing with nginx_enable_ssl.