Explicitly set the draft-origin location for all apps
Apps use Plek to find where to talk to each other, but it's also used
to configure publicly facing things like the root of the website. The
draft-origin is a bit of a grey area in between, as many applications
link to it (to show users a preview of draft-content) rather than
communicating with it internally.

With the AWS deployment of GOV.UK, this should fix the issue where
apps are using the internal domain for the draft-origin.