Make use of checkoutDependent for govuk-content-schemas
This uses the recently introduced checkoutDependent function for cloning
content schemas as part of govuk_jenkins.

It also introduces something that I was going to raise as a separate PR
but seems minor enough to be here. Rather than using a relative path
"tmp/govuk-content-schemas" as the value of the constant this uses the
absolute path which allows schemas to be used more in dependent repos
easier and as far as I'm aware has no downsides.