Restart procfile workers on environment variable changes
Previously, if there was a change to environment variables in a
deployment the worker processes would not be able to access the
new values. Sidekiq automatically reloads the workers so they are
running with the latest application code, but the Sidekiq process
itself would still be running with old environment variables.

This uses the Puppet arrow notation:
https://puppet.com/docs/puppet/5.3/lang_relationships.html#syntax-chaining-arrows