Remove conditional from the broken link checker job
The indentation and position of the command is wrong. Rather than
trying to figure out how to layout the erb, lets try removing the
conditional and changing the behaviour not to specify a machine. I've
had a quick look at the rake task being run, and it doesn't look to be
machine dependant.