Build docker images in parallel in govuk_jenkins
Building docker images can take quite a substantial amount of time in an
applications build process - regularly 4 or 5 minutes - when there isn't a
recent cache available. This utilises the [parallel][1] command in
jenkins to run the docker parts of the build in a separate thread to the
rest of the build.

A disadvantage to running things in parallel is that the output is less
easy to read as things get mixed together. To try combat this I've set
docker build to run in quiet so it only has minimal output,
unfortunately there is not an option to pass the quiet arg to docker
push tag (we could get around this by running the push tag command via
shell but that seems a bit overkill).

This sets a project that doesn't have a docker file or is running a
schema test to run just the non docker tasks in the main thread.

What was rather tricky was naming for the functions and the labels for
the output. I went with the rather uninspiring `nonDockerBuildTasks` and
`dockerBuildTasks` function names with "build" and "docker" labels for
lack of a better idea.

[1]: https://www.cloudbees.com/blog/parallelism-and-distributed-builds-jenkins