AWS set deploy user to manage assets dir
On VMware files in the uploads directory are owned by the 'assets'
user. This is required to facilitate SSH key sharing between the
asset-master and asset-slave machines. In the asset-master machine,
all file processing and file sync cron jobs belong to the 'assets'
user; data syncs from other environments are executed with the 'assets'
user, and files written on the directory from application processes
are owned by 'assets' via NFS settings.

On AWS we are not using NFS, we are using EFS instead. This is a problem
because applications are creating files with the app user 'deploy'. Then
file processing and environment sync jobs are failing because the 'assets'
user that executes those processes can't remove files created by 'deploy'.

On AWS we don't have the requirement of data sync between an asset master
and slave machines, so there is not reason for not using the 'deploy'
user to manage the data processes and own the uploads directory. This commit
updates the directories and cron jobs ownership on AWS to use 'deploy'.