Remove the AWS conditional from the Plek config
For the draft frontend nodes.

As the removed comment says, this was done when the
PLEK_SERVICE_..._URI environment variables were set when using
AWS. This is no longer the case, and these environment variables need
setting for apps on the draft frontend boxes, to make sure they use
the right URLs (e.g. search not draft-search).