AWS set UID/GID for deploy user on EFS dep boxes
Following the work done in https://github.com/alphagov/govuk-puppet/pull/7097
we need to ensure the UID/GID of the 'deploy' user is the same on all the machines
that share the EFS volume.

Without this setting, files can appear as belonging to different
users, for instance, currently the whitehall-backend machine uses the deploy
user (1058) which == jonathanhallam on asset-master.

At the moment we are forcing this change on asset-master, backend
and whitehall-backend machines, that are the only ones sharing
the volume on Integration AWS. For Staging and Production, we should
probably have the same setting everywhere, and at some point
change it in the rest of Integration too.