Add Content Audit Tool
The Audit Tool has been a sub-application of the Content Performance
Manager. We are splitting the apps as the Audit Tool won't be developed,
but we are going to be iterating the Content Performance Manager
substantially.

The Audit Tool is in private beta with users, so we need to ensure that
they still have access to it. Deploying to its own setup is the best
option now.