Fix Signon URI for authenticating-proxy in development VM
In the development VM, none of the apps (including signon) are running
with SSL enabled and so running the draft stack (which includes
authenticating_proxy) doesn't work unless the `PLEK_SERVICE_SIGNON_URI`
environment variable uses the `http` scheme vs `https`, i.e. it should
be `http://signon.dev.gov.uk`.

I had hoped to use the existing `nginx_enable_ssl` hieradata vaule which
is set to `false` in `hieradata/development.yaml`. However, it is also
set to `false` in `hieradata_aws/common.yaml` with the following
comment: "We are going to terminate SSL using ELBs". So I'm not
confident this would give the correct behaviour in AWS environments.

Note that I was initially confused why this needed to be set at all,
because I thought that the default value provided by Plek would be the
right value. However, reading this commit note [1], it seems that on
draft cache machines, Plek is configured to prefix hosts with `draft-`,
which is why we need to override it with the value from the env var.

[1]:
https://github.com/alphagov/govuk-puppet/commit/60ac13318618f9d330b023a961a74c26d6363439