Further removes @fields to fix nginx graphs
Following a PR to remove "@fields" the nginx template for logs (seen here: https://github.com/alphagov/govuk-puppet/pull/6658/files), a bunch of graphs in grafana and graphite have stopped reporting data (from 11th of Jan 2018).

This PR aims to remove the mentions of "@fields" in order to bring back the graphs.