Jenkins job copy attachments from Stg to AWS
We had to rebuild the asset-master instance on AWS to fix permissions
in the shared directory, and now we need to sync all the attachments.

Given the high volume of data to sync, we are going to run a single job
to copy the data from Staging to AWS, that will probably take a couple
of days.

The job is configured on Jenkins in Staging to avoid load/traffic increase
on production.