Move default Jenkins Jobs to common.yaml
The motivation behind this change is that we need to overwrite the
settings for the `jenkins` node. While default configuration lives in
the class based `jenkins.yaml`, we would have to create a new hierarchy
for `class/staging/jenkins.yaml` to overwrite it. This isn't really
required and just adds more places to look to find different data.

If we move the default configuration to `common.yaml`, then we can
overwrite it using `staging.yaml`, which is clearer to understand.

This commit also removes some unneeded hieradata related to deploying
stuff with vCloud Tools.