Remove private-asset-manager vhost alias
Also remove the associated environment variable which made the vhost
name available to the Asset Manager Rails app.

This virtual host alias was never used and the related functionality in
Asset Manager was removed in this PR [1].

Note that we will need to open a separate PR to remove the environment
variable definition once this PR has been merged and applied in all
environments.

[1]: https://github.com/alphagov/asset-manager/pull/426