DRY up config in static_extra_nginx_config
Note the addition to puppet-lint/plugins/check_hiera.rb. This allows me
to share the array of asset_manager_routes between the assets-origin and
static nginx config without triggering the following lint error:

    ERROR: explicit hiera() lookup for 'router::assets_origin::asset_manager_routes'

I tried using `alias`[1] to share the array of asset_manager_routes but
that doesn't seem to be available in the version of hiera that we're
using.

[1]: https://puppet.com/docs/hiera/3.3/variables.html#the-alias-lookup-function