Include draft-assets vhost in all environments
This mirrors the way we're configuring assets-origin; by including it in
this nginx.pp file.

This means that the draft-assets virtual host config will be generated
on the cache and draft-cache servers. We'll use DNS to ensure that
requests are routed to/served by the draft-cache servers.

I think ideally the draft-assets nginx config would only be on the
draft-cache servers but we already have a similar problem with the
assets-origin config being on both the cache and draft-cache servers so
I'm not going to do anything about this now.