Use env var to override Plek for draft-assets in AWS environments
The Asset Manager app needs to look up 'draft-assets' in order to decide
whether to authenticate requests and to be able to redirect requests for
draft assets. Thus the URI returned from `Plek#find` for 'draft-assets'
needs to be on the public-facing domain and not the domain for internal
services which is the default.

This is quite similar to the setting of `PLEK_SERVICE_DRAFT_ORIGIN_URI`
in `govuk::deploy::config` [1], although in this case it's specific to
the Asset Manager app only.

The aim is for `Plek#find` to return the following values for
'draft-assets':

* Development VM: 'http://draft-assets.dev.gov.uk'
* Integration (AWS): 'https://draft-assets.integration.publishing.service.gov.uk'
* Staging: 'https://draft-assets.staging.publishing.service.gov.uk'
* Production 'https://draft-assets.publishing.service.gov.uk'

All except the value for integration would've been correct before this
change.

[1]: https://github.com/alphagov/govuk-puppet/blob/19a1a860d2f0c5f13149e21e5e8a60ced9f19b2e/modules/govuk/manifests/deploy/config.pp#L107-L110