Move asset-manager nginx config to its own file
This is more consistent with the equivalent nginx config for static
which is stored in static_extra_nginx_config.conf.erb.

I ran `govuk_puppet` on the dev-vm before and after making this change.
The only differences in the asset-manager.dev.gov.uk config file
before/after applying this change are non-significant whitespace.