Add signon-sidekiq process to dev VM
I've tried to follow the pattern used by other apps, but some use a
"-worker" suffix while other use "-sidekiq". The latter seems to be in
the majority, so I've gone with that.

I'm not 100% sure the sidekiq process should depend on the signon app by
default, but I assume it should. I needed it running in order to test
the migration in this Signon PR [1].

I was also slightly confused by the fact that Signon's
`config/sidekiq.yml` defines a log file rather than outputting to
standard out. However, I haven't changed that in this PR.

[1]: https://github.com/alphagov/signon/pull/672