Increase intervals of mongodb backups in AWS
Full dumps of MongoDB are taking longer than 15 minutes, and the lock
does not wait for a clean slot so the cron job does not run. The "daily"
backup works on the basis of what time the job is started to say if it
is "daily" or not; so if a job overruns, then a daily job will never
run.

Increasing the interval will mean that the jobs should run and complete
every half an hour, without any missing jobs due to weird timings.