Skip master releases for schema tests
Although it is unusual for schema tests to run against master it is
still do-able and they were previously breaking. To execute the block
there is an expectation that a docker image is created, and this skipped
on a schema test.

We also don't need to do the various releases if we are running as a
schema test so this will be skipped.