Define unicorn workers for local-links-manager in hiera
We've upgraded local-links-manager to configure unicorn via
govuk_app_config so we want to use the env var to define the number of
worker processes now.

We set the workers to 4 because local-links-manager has had 4 unicorn
workers since alphagov/local-links-manager#52

We also have to set the value as a string in hieradata as otherwise
the validate_string function in govuk::app::envvar will complain.