Add unicorn_worker_processes to govuk::app::config
And to govuk::app so it can be passed down.  This will set an env var
that can be read by govuk_app_config and used to configure the number
of worker processes used by unicorn.  The default value is undef and
this means don't set the env var.  If the env var is not set
govuk_app_config will use it's own default value (currently 2).