Define unicorn workers for signon in hiera
We've upgraded signon to configure unicorn via govuk_app_config so we
want to use the env var to define the number of worker processes now.

We set the workers to 4 because signon has had 4 unicorn workers since
alphagov/signon#523

We also have to set the value as a string in hieradata as otherwise
the validate_string function in govuk::app::envvar will complain.