Unmount & remove /data/master-uploads from asset slaves
This was made redundant in this PR [1] when the only remaining use of
the NFS mount was removed. The comment above this `mount` resource
definition is encouraging in this respect.

Asset Manager assets are no longer permanently saved to the NFS mount
which is why the cron job copying them to the slaves was removed in [1].

Whitehall assets are "pushed" to the slaves by the
`process-incoming-files`, `process-draft-incoming-files` &
`copy-attachments-to-slaves` cron jobs using `rsync` and thus the NFS
mount is not needed on the asset slaves.

As far as I can tell, neither the Publisher (`/data/uploads/publisher`)
nor Support API assets (`/data/uploads/support-api`) require the NFS
mount on the slaves.

[1]: https://github.com/alphagov/govuk-puppet/pull/7016