Restrict asset requests that end up in Whitehall
As of the 1 Feb 2018 the only asset requests being served by Whitehall
are for attachments and HMRC. I confirmed this by using the following
query in Kibana:

    request:/GET \/government\/uploads\/.*/
    AND application:/whitehall.*/
    AND status:200
    -attachment_data
    -hmrc

I've also looked at the assets stored in the Whitehall directory on the
NFS mount and confirmed that the only files remaining are attachments
and the special case HMRC files.

Note that the HMRC assets URLs use www.gov.uk instead of
assets.publishing.service.gov.uk. I've added an explicit route for them
here in case anyone was relying on being able to access them via the
assets host.

Moving the '/government/uploads/' catch-all route means that all
non-attachment and non-HMRC asset requests are now sent to asset-manager
(via static). This is preferable to sending them to Whitehall as, since
https://github.com/alphagov/whitehall/pull/3763, sending them to
Whitehall would result in an infinite redirect loop.