Validate that .ruby-version and Dockerfile have matching version
It is important that the Dockerfile matches the .ruby-version as
we use the Dockerfile as part of the publishing-e2e-tests.  If the
.ruby-version gets bumped without the Dockerfile then the
publishing-e2e-tests could give you a false negative that the
upgrade is non-breaking.

There isn't a way to automatically pull the .ruby-version into the
Dockerfile. So in place of that add a CI check which validates they
are at parity.

We ignore any information after the - from both sides to allow for
minor variations.  2.0.0-p648 == 2.0.0-alpine