Add repo for ValeLint
ValeLint is a super useful tool that I'd like to set up in our CI
environment to do linting across our documentation.

We can begin with govuk-developer-docs, but it may be worth running
against anything in .md files and adding to the default Jenkinslib.

This creates a repository so I can add the package available for
install.