Disable transition logs in AWS
Transition Logs is a thing which runs some analysis on the logs received
from CDN and commits the results to Github, which is then picked up by
the Transition app.

While we're testing we do not want to risk accidentally copying
credentials over and committing incorrect data to the Github repository,
so we should disable it entirely in AWS until we're ready to migrate the
process to AWS.

This should not stop us testing receiving CDN logs from Fastly to AWS,
which is the principle function of this instance.