Add Integration specific jenkins.yaml
I believe Puppet deploys have broken in Integration since this PR:
https://github.com/alphagov/govuk-secrets/pull/245

This is because we do not have a directory called 'integration' which is
causing the job to fail.

We should create an Integration directory to copy, but I can only check
it in if I have a file inside it, so I'm using the Jenkins specific
theme for Integration to do this.