Add ValeLint to CI agents
ValeLint is a tool that lints documentation: https://github.com/ValeLint/vale

We tried it on a pre-commit hook for govuk-developer-docs, but since
this relies on trusting a someone has installed on govuk-developer-docs
it is a less reliable way to ensure all our documentation is correctly
linted.

To begin with we can start with a low threshold for error, but can catch
fully incorrect grammer such as repeated words.

This will also need to be configured in the Jenkinsfile.

In the future we could potentially configure the jenkinslib to default
check any markdown files and ensure that we have well-written
documentation.