Don't set PLEK_HOSTNAME_PREFIX on the draft_cache nodes
The only applications which run on these nodes are
authenticating-proxy, router and router-api. The only one of these
that uses Plek is authenticating-proxy, and the only thing it uses
Plek for is to find Signon.

There is no draft Signon, so previously this was overridden in the
authenticating_proxy class. I removed that override breaking
authenticating-proxy, but rather than putting it back, I think it
makes more sense to remove the use of PLEK_HOSTNAME_PREFIX for the
draft_cache nodes. Less configuration is probably better, especially
if I'm correct in saying that the setting is redundant.