Add release app and slack build params to CI deploy trigger job
If we don't send these build parameters in the trigger then 
jenkins doesn't default them and so the deployment won't try
to talk to the release app or to slack.  It doesn't actually talk
to slack anyway, because there's code that only does it for
staging or production, but that might change and we wouldn't
want to have to edit this script again to enable it.