Install libffi-dev with all rbenv installs
This has been added to govuk_rbenv to bring it as close as possible to
installing for all environments which have Ruby. It is to resolve an
issue we have where a version of the ffi gem is blocking deployments on
a number of apps due to being able to install this gem, however as the
libffi-dev package is installed on CI these are aready merged in quite a
few places.

Therefore it felt like the path of least pain to install this package in
all ruby installations so that gems requiring it would install
consistently.

It looks like this issue should be resolved soon by this commit:
https://github.com/ffi/ffi/commit/bfc330a3c6cbeedea72a0134d603085b54953252
however the project does specify that you need libffi-dev installed
(however that could be just for developing) so there is a fixme note to
check and remove this once ffi has been updated.