Add global shared library to Jenkins
This is related to
https://github.com/alphagov/govuk-jenkinslib/commit/c2e84ed07a17d2afd17d090bf4a87b0d62efeef9.

This sets the global option to enable loading in a shared library from a
repository. Pass in a hash of the libraries you wish to load, and then
you should be able to load the libraries in using a Jenkinsfile.

I used an embedded hash within the manifest because at the moment it's a
single library. In the future it may be easier to create a default set
of values, and pass in a hash from hiera.

There are a few advantages over loading in an external shared library,
including keeping the library in it's own version control, being able to
potentially add unit tests, because able to explictly call a job from a
branch of the library to test changes and making the code more portable
and visible.

Implicit load is set to false which means the library must be explicitly
loaded in through a Jenkinsfile.