Sets email api delivery threshold as string
Following the deployment of this PR : https://github.com/alphagov/govuk-puppet/commit/fc142c9fedb5f187160c1a8d4774e833e6aafacf there have been a number of errors regarding the fact that the threshold for the email api (which was set to 3000) is a Fixnum and not a String.

```
Could not retrieve catalog from remote server: Error 400 on SERVER: 3000 is not a string.  It looks to be a Fixnum ...
```