Start manuals and specialist publisher in a consistent manner
These apps were using foreman to boot up which can cause surprises in a
dev environment with the work that that has been going on with the
procfiles in apps. This changes them to use rails server as is
consistent with other apps.

See: https://github.com/alphagov/specialist-publisher/pull/1167
and https://github.com/alphagov/manuals-publisher/pull/1291 for more
info.