Better MFA / session handling
Ask for a MFA token on the command line, and then store the credentials
in a file. The `assume-role` command returns an expiration time, so
check this with a 5 minute margin of error. If the token has expired
or is about to expire, prompt the user for their MFA token again.

This has the added benefit of subsequent runs of this script not
requiring a new MFA token if the STS session is still valid.