Disable checks for Router on Docker
The UFW rule and healthcheck are provided by the default included class
govuk::node::s_cache. Rather than try to ensure this is not included
(which becomes complicated quickly), just ensure the resources that
clash (UFW and Icinga checks) are not enabled.

This is a short-term change to work enable unblocking work and this
class will not be used in the long-term, so is the safest place to make
these changes.