Increase icinga threshold for whitehall frontend CPU alert
We are getting lots of alerts from whitehall frontend when the CPU usage goes
pass the current thresholds, however this server has 8 CPU and as such the
default thresholds don't make much sense. I have set this up so that staging
and prod are configures with higher thresholds.

I have initentionally not updated AWS data as this will need to be reviewed
once the new machine are started and we are better off getting false positives
on these alarms (i.e. leave it with the lower threshold to begin with)