Use external GOV.UK Jenkins library
This uses the library defined in [the external Jenkins library](https://github.com/alphagov/govuk-jenkinslib)
hosted in an external repository.

The library is not implicitly loaded. The following needs
to be defined in each `Jenkinsfile`:

`library("govuk")`

It is possible to load different branches of the upstream library by
defining the branch or SHA:

`library("govuk@my-new-branch")`