Disable email-alert-api procfile worker on backend boxes
This is a step in switching the boxes for Email Alert API from
email-alert-api class to backend class.

This disable procfile workers on backend machines and thus the procfiles
will only be running on the email-alert-api machines. The idea of this
is that we can run email-alert-api as web service across both machines
but only have the workers on the new boxes, thereby limiting the db
connections used (as if we run all the processes on all boxes we blow
the postgres connection limit)