Add a new metric to the Topic Taxonomy dashboard
We are now recording the average taxons per tagged content item, so
show this on the dashboard. Currently the data is only available in
the integration environment.

To keep the dashboard small, move the change information for the
average tagging depth and average taxons per tagged content item
metrics in to a row which is hidden by default.